get rid of the warnings

+++ cpucounters.cpp
@@ -5193,19 +5193,19 @@ void PCM::programPCIeCounters(const PCM::PCIeEventCode event_, const uint32 tid_
                 switch(q_)
                 {
                     case PRQ:
-                        umask |= SKX_CHA_TOR_INSERTS_UMASK_PRQ(1);
+                        umask |= (uint64)(SKX_CHA_TOR_INSERTS_UMASK_PRQ(1));
                         break;
                     case IRQ:
-                        umask |= SKX_CHA_TOR_INSERTS_UMASK_IRQ(1);
+                        umask |= (uint64)(SKX_CHA_TOR_INSERTS_UMASK_IRQ(1));
                         break;
                 }
                 switch(miss_)
                 {
                     case 0:
-                        umask |= SKX_CHA_TOR_INSERTS_UMASK_HIT(1);
+                        umask |= (uint64)(SKX_CHA_TOR_INSERTS_UMASK_HIT(1));
                         break;
                     case 1:
-                        umask |= SKX_CHA_TOR_INSERTS_UMASK_MISS(1);
+                        umask |= (uint64)(SKX_CHA_TOR_INSERTS_UMASK_MISS(1));
                         break;
                 }
